Move `get_downloaded_document` so it can be reused
